loyha deyarli tugadi

+++ server/src/modules/category/category.controller.js
@@ -1,4 +1,6 @@
 import categoryService from "./category.service.js";
+import { isValidObjectId } from "mongoose";
+import { BaseException } from "../../exception/base.exception.js";
 
 class userController{
     #_categoryService
@@ -9,7 +11,10 @@ class userController{
     getAllCategory = async (req,res,next) => {
         try {
             const data = await this.#_categoryService.getAllCategory();
-            res.send(data);
+            res.status(200).send({
+                message:"success",
+                data:data
+            })
         } catch (error) {
             next(error)
         }
@@ -18,8 +23,19 @@ class userController{
     getCategory = async (req,res,next) => {
         try {
             const id = req.params.id;
+            
+            if (!(isValidObjectId(id))) {
+                throw new BaseException(`given id: ${id} is not valid`,400)
+            }
+
             const data = await this.#_categoryService.getCategory(id);
-            res.send(data);
+            if(data==null){
+                throw new BaseException("Category not found",404)
+            }
+            res.status(200).send({
+                message:"success",
+                data:data,
+            })
         } catch (error) {
             next(error);
         }
@@ -29,7 +45,10 @@ class userController{
         try {
             const {name,goods,shelve}= req.body;
             const data = await this.#_categoryService.createCategory(name,shelve);
-            res.send(data)
+            res.status(200).send({
+                message:"success",
+                data:data
+            })
         } catch (error) {
             next(error);
         }
@@ -39,8 +58,19 @@ class userController{
         try {  
             const {name,goods,shelve}= req.body;
             const id = req.params.id;
+            
+            if (!(isValidObjectId(id))) {
+                throw new BaseException(`given id: ${id} is not valid`,400)
+            }
+
             const data = await this.#_categoryService.updateCategory(id,name,shelve);
-            res.send(data)
+            if(data.error){
+                throw new BaseException(data.error,data.status)
+            }
+            res.status(200).send({
+                message:"success",
+                data:data
+            })
         } catch (error) {
             next(error)
         }
@@ -49,8 +79,19 @@ class userController{
     deleteCategory = async (req,res,next) => {
         try {
             const id = req.params.id;
+            
+            if (!(isValidObjectId(id))) {
+                throw new BaseException(`given id: ${id} is not valid`,400)
+            }
+
             const data = await this.#_categoryService.deleteCategory(id);
-            res.send(data)
+            if(data.error){
+                throw new BaseException(data.error,data.status)
+            }
+            res.status(200).send({
+                message:"success",
+                data:data
+            })
         } catch (error) {
             next(error);
         }
