+++ server/src/modules/warehouse/warehouse.controller.js
@@ -1,4 +1,6 @@
-import warehouseService from "./warehouse.service"
+import warehouseService from "./warehouse.service.js"
+import { isValidObjectId } from "mongoose";
+import { BaseException } from "../../exception/base.exception.js";
 
 
 class WarehouseController{
@@ -9,7 +11,10 @@ class WarehouseController{
     getAllWarehouses = async (req,res,next) => {
         try {
             const data = await this.#_warehouseService.getAllWarehouses();
-            res.send(data)
+            res.status(200).send({
+                message:"success",
+                data:data
+            })
         } catch (error) {
             next(error)
         }
@@ -18,8 +23,16 @@ class WarehouseController{
     getWarehouse = async (req,res,next) => {
         try {
             const id = req.params.id;
+            
+            if (!(isValidObjectId(id))) {
+                throw new BaseException(`given id: ${id} is not valid`,400)
+            }
+
             const data = await this.#_warehouseService.getWarehouse(id);
-            res.send(data)
+            res.status(200).send({
+                message:"success",
+                data:data
+            })
         } catch (error) {
             next(error)  
         }
@@ -27,9 +40,15 @@ class WarehouseController{
 
     createWarehouse = async (req,res,next) => {
         try {
-            const {name,type,owner} = req.body;
-            const data = await this.#_warehouseService.createWarehouse(name,type,owner);
-            res.send(data)
+            const {name,type,owner,category} = req.body;
+            const data = await this.#_warehouseService.createWarehouse(name,type,owner,category);
+            if(data.error){
+                throw new BaseException(data.error,data.status);
+            }
+            res.status(200).send({
+                message:"Successfully created!",
+                data:data
+            })
         } catch (error) {
             next(error)  
         }
@@ -38,9 +57,20 @@ class WarehouseController{
     updateWarehouse = async (req,res,next) => {
         try {
             const id = req.params.id;
-            const {name,type,owner} = req.body;
+            
+            if (!(isValidObjectId(id))) {
+                throw new BaseException(`given id: ${id} is not valid`,400)
+            }
+
+            const {name,type,owner,category} = req.body;
             const data = await this.#_warehouseService.updateWarehouse(id,name,type,owner);
-            res.send(data)
+            if(data.error){
+                throw new BaseException(data.error,data.status);
+            }
+            res.status(200).send({
+                message:"Successfully updated!",
+                data:data
+            })
         } catch (error) {
             next(error)
         }
@@ -49,8 +79,19 @@ class WarehouseController{
     deleteWarehouse = async (req,res,next) => {
         try {
             const id = req.params.id;
+            
+            if (!(isValidObjectId(id))) {
+                throw new BaseException(`given id: ${id} is not valid`,400)
+            }
+
             const data = await this.#_warehouseService.deleteWarehouse(id);
-            res.send(data)
+            if(data.error){
+                throw new BaseException(data.error,data.status);
+            }
+            res.status(200).send({
+                message:"Successfully deleted!",
+                data:data
+            })
         } catch (error) {
             next(error)
         }
